优化了取数流程
Controller直接调用IDomainFacotry内容，减少了一个中间环节，结构更清晰。

+++ PHMS2/Models/ViewModels/InPatientReporter/InPatientAverageAntibioticCostRate.cs
@@ -19,7 +19,7 @@ namespace PHMS2.Models.ViewModels.InPatientReporter
         {
             get
             {
-                return this.TotalAntibioticPerson != 0 ? Decimal.Round(this.TotalAntibioticCost  / (Decimal) this.TotalAntibioticPerson, 2) : 0;
+                return this.TotalAntibioticPerson != 0 ? Decimal.Round((Decimal)this.TotalAntibioticCost  / (Decimal) this.TotalAntibioticPerson, 2) : 0;
             }
         }
     }
